feat: EAS-105 integration with devcontainers (#17)
* added vscode workspace

* modularized api startup to allow swagger schema generation

* build: updated tsconfig

* feat: monorepo with dev containers

* fix: removed webapp devcontainer-specific settings

* fix: webapp entrypoint

* fix: schema generation path
