+++ .devcontainer/webapp/devcontainer.json
@@ -1,0 +1,12 @@
+{
+  "name": "Easymotion Webapp",
+  "dockerComposeFile": ["../../docker-compose.yml"],
+  "service": "webapp",
+  "shutdownAction": "none",
+  "workspaceFolder": "/workspace/webapp",
+  "customizations": {
+    "vscode": {
+      "extensions": ["esbenp.prettier-vscode", "dbaeumer.vscode-eslint"]
+    }
+  }
+}
