+++ api/.vscode/settings.json
@@ -1,0 +1,16 @@
+{
+  "jest.shell": "/bin/sh",
+  "jest.jestCommandLine": "",
+  "editor.formatOnSave": true,
+  "editor.defaultFormatter": "esbenp.prettier-vscode",
+  "editor.tabSize": 2,
+  "editor.insertSpaces": true,
+  "editor.detectIndentation": false,
+  "typescript.tsserver.experimental.enableProjectDiagnostics": true,
+
+  "eslint.workingDirectories": [{ "mode": "auto" }],
+  "eslint.validate": ["javascript", "typescript"],
+
+  "prettier.useEditorConfig": true,
+  "prettier.configPath": ""
+}
