+++ api/Dockerfile.dev
@@ -1,0 +1,9 @@
+FROM node:20-alpine
+
+RUN apk add --no-cache libc6-compat
+RUN apk add --no-cache openssl
+RUN apk add --no-cache curl
+
+WORKDIR /workspace/api
+
+ENV NODE_ENV=development
