+++ api/package.json
@@ -23,7 +23,7 @@
     "prisma:migrate-dev": "prisma migrate dev",
     "prisma:migrate-prod": "prisma migrate deploy",
     "prisma:gen-client": "prisma generate client",
-    "swagger:generate": "ts-node ./src/generate-schema.ts"
+    "swagger:generate": "nest start --entryFile src/generate-schema"
   },
   "dependencies": {
     "@nestjs/common": "^10.0.0",
