+++ api/src/bootstrap.ts
@@ -1,0 +1,58 @@
+import { HttpAdapterHost, NestFactory, Reflector } from '@nestjs/core';
+import { AppModule } from './app.module';
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { ClassSerializerInterceptor, ValidationPipe } from '@nestjs/common';
+import {
+  API_DESCRIPTION,
+  API_TITLE,
+  API_VERSION,
+} from './common/constants/swagger.constants';
+import { PrismaClientExceptionFilter } from 'nestjs-prisma';
+import { ConfigService } from '@nestjs/config';
+
+export async function bootstrap() {
+  const app = await NestFactory.create(AppModule, {
+    logger: ['error', 'warn', 'log', 'debug', 'verbose'],
+  });
+
+  const configService = app.get(ConfigService);
+
+  const { httpAdapter } = app.get(HttpAdapterHost);
+  app.useGlobalFilters(new PrismaClientExceptionFilter(httpAdapter));
+
+  const corsOrigin = configService.get<string>('API_CORS_ORIGIN');
+
+  app.enableCors({
+    origin: corsOrigin ?? 'https://easymotion.devlocal',
+    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
+    credentials: true, // Allows cookies or auth headers
+  });
+
+  app.useGlobalPipes(
+    new ValidationPipe({
+      transform: true,
+      whitelist: true,
+      forbidNonWhitelisted: false,
+    }),
+  );
+
+  app.useGlobalInterceptors(new ClassSerializerInterceptor(app.get(Reflector)));
+
+  return app;
+}
+
+export function setupSwagger(app) {
+  const config = new DocumentBuilder()
+    .setTitle(API_TITLE)
+    .setDescription(API_DESCRIPTION)
+    .setVersion(API_VERSION)
+    .build();
+
+  const document = SwaggerModule.createDocument(app, config);
+
+  SwaggerModule.setup('swagger', app, document, {
+    jsonDocumentUrl: 'swagger/json',
+  });
+
+  return document;
+}
