+++ webapp/Dockerfile
@@ -1,14 +1,4 @@
-# Setup dev image
-
-FROM node:20-alpine as dev 
-
-WORKDIR /app
-
-ENV NODE_ENV development
-
-# Build the app
-
-FROM node:20-alpine as build
+FROM node:20-alpine AS build
 
 WORKDIR /app
 
@@ -25,12 +15,12 @@ USER node
 
 # Prepare production image
 
-FROM node:20-alpine as prod
+FROM node:20-alpine AS prod
 
 WORKDIR /app
 RUN apk add --no-cache libc6-compat
 
-ENV NODE_ENV production
+ENV NODE_ENV=production
 
 RUN addgroup --system --gid 1001 node || true
 RUN adduser --system --uid 1001 node || true
