+++ webapp/Dockerfile.dev
@@ -1,0 +1,5 @@
+FROM node:20-alpine
+
+WORKDIR /workspace/webapp
+
+ENV NODE_ENV=development
