+++ webapp/package.json
@@ -7,7 +7,7 @@
     "dev": "NODE_TLS_REJECT_UNAUTHORIZED=0 vite --host 0.0.0.0",
     "build": "npm run generate-client && tsc -b && vite build",
     "lint": "eslint .",
-    "generate-client": "swagger-typescript-api -p ./api-schema/schema.json -o ./client --modular",
+    "generate-client": "swagger-typescript-api -p /workspace/webapp/api-schema/schema.json -o ./client --modular",
     "preview": "vite preview",
     "test": "vitest",
     "test:watch": "vitest --watch",
